requirements.txt
#sudo apt-get install python3.9-dev
geopandas==1.0.1
pandas==2.2.3
numpy>=1.25
shapely==2.0.6  # 1.7.1
Rtree==1.3.0
requests==2.32.3
rasterio==1.3.11
xarray==2024.7.0
rioxarray==0.13.4
cdsapi==0.7.4
gdal==3.8 # Matching GDAL system library which is 3.8.4 (so, update if the system library is also changed)
netcdf4==1.7.2
esgf-pyclient==0.3.1
matplotlib==3.9.4
pydap==3.5
pysheds==0.4
cfgrib==0.9.15
pyvista==0.44.2
fire==0.7.0
json5==0.10.0
topopyscale
gitpython==3.1.44
gspread
google-auth